requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Jinja2==2.11.2
PyYAML==5.3.1
requests==2.25.1
SQLAlchemy==1.3.23
toml==0.10.2
twilio==6.52.0
twitter==1.18.0
urllib3==1.26.2
